@@ -1,12 +1,17 @@
 package gaoyang.biometricdemo;
 
+import android.Manifest;
+import android.content.pm.PackageManager;
 import android.os.Build;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
 import android.widget.TextView;
 import android.widget.Toast;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.core.app.ActivityCompat;
+import androidx.core.content.ContextCompat;
 
 import com.nestia.biometriclib.BiometricPromptManager;
 
@@ -15,22 +20,45 @@ public class MainActivity extends AppCompatActivity {
     private TextView mTextView;
     private Button mButton;
     private BiometricPromptManager mManager;
+    private String[] permissions = {Manifest.permission.WRITE_EXTERNAL_STORAGE,
+            Manifest.permission.USE_FINGERPRINT, Manifest.permission.USE_BIOMETRIC, Manifest.permission.ACCESS_LOCATION_EXTRA_COMMANDS};
+
+    private void initPermission() {
+        // 版本判断。当手机系统大于 23 时，才有必要去判断权限是否获取
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+            // 检查该权限是否已经获取
+            int i = ContextCompat.checkSelfPermission(getApplicationContext(), permissions[0]);
+            int l = ContextCompat.checkSelfPermission(getApplicationContext(), permissions[1]);
+            int m = ContextCompat.checkSelfPermission(getApplicationContext(), permissions[2]);
+            int n = ContextCompat.checkSelfPermission(getApplicationContext(), permissions[3]);
+            // 权限是否已经 授权 GRANTED---授权  DINIED---拒绝
+            if (i != PackageManager.PERMISSION_GRANTED || l != PackageManager.PERMISSION_GRANTED || m != PackageManager.PERMISSION_GRANTED ||
+                    n != PackageManager.PERMISSION_GRANTED) {
+                // 如果没有授予该权限，就去提示用户请求
+                ActivityCompat.requestPermissions(MainActivity.this, permissions, 321);
+            }
+        }
+    }
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
+
+        initPermission();
+
         mTextView = findViewById(R.id.text_view);
         mButton = findViewById(R.id.button);
 
         mManager = BiometricPromptManager.from(this);
         StringBuilder stringBuilder = new StringBuilder();
-        stringBuilder.append("SDK version is "+ Build.VERSION.SDK_INT);
+        stringBuilder.append("SDK version is " + Build.VERSION.SDK_INT);
         stringBuilder.append("\n");
-        stringBuilder.append("isHardwareDetected : "+mManager.isHardwareDetected());
+        stringBuilder.append("isHardwareDetected : " + mManager.isHardwareDetected());
         stringBuilder.append("\n");
-        stringBuilder.append("hasEnrolledFingerprints : "+mManager.hasEnrolledFingerprints());
+        stringBuilder.append("hasEnrolledFingerprints : " + mManager.hasEnrolledFingerprints());
         stringBuilder.append("\n");
-        stringBuilder.append("isKeyguardSecure : "+mManager.isKeyguardSecure());
+        stringBuilder.append("isKeyguardSecure : " + mManager.isKeyguardSecure());
         stringBuilder.append("\n");
 
         mTextView.setText(stringBuilder.toString());
