@@ -2,8 +2,8 @@ package com.nestia.biometriclib;
 
 import android.content.Context;
 import android.content.SharedPreferences;
-import android.support.annotation.NonNull;
 
+import androidx.annotation.NonNull;
 
 public class SPUtils {
 
